Move auth package to `ui/login`
