+++ frontend/app/src/main/java/com/example/frontend/ui/BottomNavigation.kt
@@ -1,0 +1,4 @@
+package com.example.frontend.ui
+
+class BottomNavigation {
+}
